python-server/requirements.txt
pydantic
openai
pymupdf
termcolor
python-dotenv